% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{assign_call}
\alias{assign_call}
\title{Allocate Data to Head}
\usage{
assign_call(data, timestamp, desired_size, features)
}
\arguments{
\item{data}{timeseries dataframe}

\item{timestamp}{POSIXct time from which to populate matrix}

\item{desired_size}{integer for size of matrix to populate}

\item{features}{timeseries features for assignment}
}
\value{
matrix populated with data
}
\description{
This function populates a matrix with the selected data at a point in time
}
