% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cerberus.R
\name{build_cerberus}
\alias{build_cerberus}
\title{Build Cerberus model}
\usage{
build_cerberus(training_data, csize = 64)
}
\arguments{
\item{training_data}{list of formatted call, response, context data}

\item{csize}{integer relative dimension of hidden layers}
}
\value{
keras Cerberus model
}
\description{
This function takes the call, response, and context heads, and combines them
into a full Cerberus model.
}
