% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{calculate_min_max}
\alias{calculate_min_max}
\title{Determine Feature Min and Max}
\usage{
calculate_min_max(data)
}
\arguments{
\item{data}{timeseries dataframe}
}
\value{
dataframe of min and max by feature
}
\description{
This function calculates the minimum and maximum of each feature in a
timeseries data set
}
