% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{cerberus_config}
\alias{cerberus_config}
\title{Cerberus Config Presets}
\usage{
cerberus_config(
  call_feature_indx,
  res_feature_indx,
  context_feature_indx,
  ts_units = "hours",
  call_ts = 2,
  call_size = 48,
  res_ts = 2,
  res_size = 12,
  context_ts = c(6),
  context_size = c(12),
  scale_range = c(0, 1)
)
}
\arguments{
\item{ts_units}{character string for unit of timeseries, referenced by size}

\item{call_ts}{integer. Resolution timesteps of ts_units for call head}

\item{call_size}{integer. Number of timesteps for call head}

\item{res_ts}{integer. Resolution timesteps of ts_units for response head}

\item{res_size}{integer. Number of timesteps for response head}

\item{context_ts}{integer array. Resolution timesteps of ts_units for context heads}

\item{context_size}{integer array. Number of timesteps for context heads}

\item{call_feature_index}{integer array of column indices for call features}

\item{res_feature_index}{integer array of column indices for response features}

\item{context_feature_index}{integer array of column indices for context features}
}
\value{
config list
}
\description{
This function provides some presets for Cerberus configuration.
The _indx arguments expect integer arrays for the column indices in the
timeseries data.
}
