% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_functions.R
\name{descale_response}
\alias{descale_response}
\title{Descale Generated Response}
\usage{
descale_response(responses, min_max_df, config, features, timestamp)
}
\arguments{
\item{responses}{generated response predictions}

\item{min_max_df}{dataframe of min and max by feature}

\item{config}{Cerberus configuration list.}

\item{features}{character array. Timeseries features used in Cerberus}

\item{timestamp}{POSIXct timestamp from which to generate predictions}
}
\value{
timeseries data frame of descaled generated response
}
\description{
This function returns a generated response from Cerberus to the original
range of the timeseries data.
}
