% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cerberus.R
\name{form_head}
\alias{form_head}
\title{Form Single Cerberus Head}
\usage{
form_head(layerinput, size, featurelen, csize = 64)
}
\arguments{
\item{layerinput}{keras layer input object}

\item{size}{integer layer input size}

\item{featurelen}{integer layer feature length}

\item{csize}{integer relative dimension of hidden layers}
}
\value{
keras Cerberus head object
}
\description{
This function forms a single head (either call, response, or context) for
Cerberus model
}
