% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_functions.R
\name{generate_prediction}
\alias{generate_prediction}
\title{Generate Response Predictions}
\usage{
generate_prediction(model, timestamp, resampled_time_series, features, config)
}
\arguments{
\item{model}{trained Cerberus keras model object}

\item{timestamp}{POSIXct timestamp from which to generate predictions}

\item{resampled_time_series}{prepared Ceberus resampled timeseries data list}

\item{features}{character array. Timeseries features used in Cerberus}

\item{config}{Cerberus configuration list.}
}
\value{
generated response predictions
}
\description{
This function uses a trained cerberus model to generatively predict responses
}
