% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{grab_train_samples}
\alias{grab_train_samples}
\title{Retrieve Training Data Samples}
\usage{
grab_train_samples(formatted_all, ratio = 0.7)
}
\arguments{
\item{formatted_all}{list of prepared Cerberus arrays.}

\item{ratio}{proportion of data to be used for training.}
}
\value{
list of training Cerberus arrays.
}
\description{
This function retrieves training data from our prepared Cerberus data list
based on the desired proportion
}
