% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cerberus.R
\name{ksize}
\alias{ksize}
\title{Determine appropriate neural network hidden layer size}
\usage{
ksize(len)
}
\arguments{
\item{len}{input size}
}
\value{
integer hidden layer size
}
\description{
Determine appropriate neural network hidden layer size
}
