% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{resample_timeseries}
\alias{resample_timeseries}
\title{Resample Timeseries to New Timestep}
\usage{
resample_timeseries(data, window_size)
}
\arguments{
\item{data}{timeseries dataframe}

\item{window_size}{integer for window size to average timeseries over}
}
\value{
resampled timeseries dataframe
}
\description{
This function resamples timeseries data by averaging over a window size
}
