% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{scale_features}
\alias{scale_features}
\title{Scale Features to Range}
\usage{
scale_features(data, min_max_df, min_range = -1, max_range = 1)
}
\arguments{
\item{data}{timeseries dataframe}

\item{min_max_df}{dataframe of min and max by feature}

\item{min_range}{minimum desired value of scaled data}

\item{max_range}{maximum desired value of scaled data}
}
\value{
scaled timeseries dataframe
}
\description{
This function scales timeseries data to a user specified range provided the
min and max of each feature
}
