% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cerberus.R
\name{train_cerberus}
\alias{train_cerberus}
\title{Train Cerberus Model}
\usage{
train_cerberus(model, training_data, epochs)
}
\arguments{
\item{model}{keras Cerberus model}

\item{training_data}{list of formatted call, response, context data for training}

\item{epochs}{desired number of epochs integer}
}
\value{
trained keras Cerberus model
}
\description{
This function takes a constructed Cerberus model and trains it to provided
timeseries data using ADAM.
}
